As pointed out by Charles Hannum, program files which are not intended to be in a shared object should not be compiled as Position Independent Code.
Attempt to fix ncurses' man page problems once and for all.  The
configure script tries to be smart, and checks the date(1) manual page
to see if it's been compressed, and, if it has been, installs ncurses
manual pages compressed.  I've removed this check, so that all pages
are installed (as far as ncurses thinks) uncompressed, and then use
the standard pkgsrc mechanisms for compressing them.