Addition of cucipop, version 1.31, to the NetBSD packages collection.
Cucipop, the Cubic Circle POP3 server, is a very small and fast server for
POP3 mail clients--far faster than UW's ipop3d.  It does not support POP2,
but it does include support for the TOP extension to POP3, Content-Length:
headers, and APOP.  It can be run from inetd or standalone.

Please note that this package is shareware.

This package is based on Todd Vierling's submission in PR pkg/4977, as well
as Jim Wise's submission in PR pkg/4910, which were put on hold until the
licence logic had been added to bsd.pkg.mk, but I've also updated this to
version 1.31, and fixed a few things in the compilation process.