Call "pkg_delete -O' on "make reinstall", to remove things not only from /var/db/pkg, but also from the pkgdb (in case someone does a "make reinstall" with FORCE_PKG_REGISTER set). Pointed out by Greg Woods.
XXX This change is only effective if one has a pkg_* tools as of 19990302
or later, but I will not bump the PKGTOOLS_REQD version up, as I do
not expect many people to run into this (and if, it's not a problem).
One should just remove the check with ${_PKGTOOLS_VER} >= 19990302
if the version's bumped beyond 19990302.