MBROLA based American English male voice for the festival speech synthesis system.
This voice provides a American English male voice using the MBROLA
synthesis method.  It uses a modified CMU lexicon for pronunciations.
Prosodic phrasing is provided by a statistically trained model using
part of speech and local distribution of breaks.  Intonation is
provided by a CART tree predicting ToBI accents and an F0 contour
generated from a model trained from natural speech.  The duration
model is also trained from data using a CART tree.

The quality of this voice is not as high as us1 and us2

This voice can be activated via (voice_us3_mbrola)

-Julian Assange <proff@iq.org>