Update MySQL to 3.22.27.
There are a bunch of bug fixes to the server.  For a complete list see
the documentation.

Also, get rid of USE_GMAKE.  The normal make works just fine.
Delete a whole bunch of patches that have been integrated into the
distribution.
Update the mit-pthreads patches to use _C_LABEL where appropriate so that
mit-pthreads works on ELF and a.out for the affected platforms.