This patch cleans up our message for unacceptable licenses:
miyu% make
===>  xephem-3.1 Unacceptable license: no-commercial-use.
        To build this package, add this line to your /etc/mk.conf:
        ACCEPTABLE_LICENSES+=no-commercial-use

To do this, we need to change the semantics of the IGNORE variable. It is
no longer taken as one variable, but rather as one variable with several
lines, seperated by whitespace. Each line will be printed on it's own.

In addition, the code for the IGNORE_FAIL has been slightly moved around
so IGNORE_SILENT can be used together with IGNORE_FAIL now. (does someone
use this?).