Import libirc into the NetBSD packages collection.
"libirc is a C-library for the IRC protocol.  It simplifies the task of
writing a bot or client.  A very simple example comes with the
package, demonstrating how it is used.

Perfect for those who wish to write bots, irc clients (especially if
you want to integrate IRC with some other Internet software).  The
programming language is C."

Note, however, that this package lacks any form of documentation.