Begin unification of install paths for java based packages.  Scheme is:
      * binaries which are not JVM dependent go in ${PREFIX}/bin

      * jar/zip/etc files for use by java programs go in
	${PREFIX}/lib/java

      * examples for java packages go in ${PREFIX}/share/examples/java

      * documentation for java packages go in ${PREFIX}/share/doc/java

JVMs go in their own directories, right now ${PREFIX}/{java,kaffe}, which
is necessary so that multiple JVMs can be installed concurrently, and so
that the miriad packages which count on ${JAVA_HOME} pointing to a JVM
install will DTRT.

Most java packages already use something similar or can be adapted easily.
Java packages such as Jakarta with large amounts of private java will need
to be addressed separately.