for NetBSD platform, pass bsd.pkg.mk ${LDFLAGS} via sendmail M4 declaration confLDOPTS.
the real problem is in sendmail side - M4 macro APPENDDEF chokes if we
pass string with "," inside.  we are okay for now, as we don't have any
APPENDDEF against confLDOPTS.  in general, we are not sure if the fix works
for everyone.

not sure if we need the similar fix for Solaris or not.

problem report From: Bernd.Ernesti@security.kpnqwest.com (Bernd Ernesti)