Redo enabling of IPv6 support in a different way.
Instead of using pkgsrc makefile magic to select between IPv6 support
and non-support do so by the NetBSD version number: on 1.5 and newer
systems, compile IPv6 support and detect kernel support for it at runtime.

This has the additional benefits of being easy to feed back and additionaly
brings IPv6 to mozilla on NetBSD even if build outside of pkgsrc.