Initial import of hdf: 	NCSA Hierarchical Data Format
HDF is a multi-object file format that facilitates the transfer of
various types of scientific data between machines and operating
systems.  HDF allows self-definitions of data content and easy
extensibility for future enhancements or compatibility with other
standard formats.  HDF includes Fortran and C calling interfaces, and
utilities to prepare raw image of data files or for use with other
NCSA software.  The HDF library contains interfaces for storing and
retrieving compressed or uncompressed 8-bit and 24-bit raster images
with palettes, n-Dimensional scientific datasets and binary tables.
An interface is also included that allows arbitray grouping of other
HDF objects.