Initial import of x48-0.4.1 into the packages collection.
Provided in PR 12134 by Jens A Nilsson <jens@subzero.campus.luth.se>

This emulator is capable of providing a faithful replication of the
HP48. In order to do so, it requires a copy of the ROM software
from YOUR calculator. In order to avoid breaking copyright laws,
and upsetting HP, you MUST BE THE PROUD OWNER OF AN HP48 before
running this program. Of course you can run this program without a
copy of the ROM software in order to write trivial machine code
programs but you will not be able to access any of the calculator
functionality.

WWW: http://home.earthlink.net/~rmmlgetz/