Update to gsl-0.8.
* What is new in gsl-0.8:

** The build process now uses the latest libtool and automake

** The library should now compile with Microsoft Visual C++

** Portable versions of the isinf, isnan and finite functions are
available as gsl_isinf(x), gsl_isnan(x) and gsl_finite(x).

** The definitions of GSL_POSINF, GSL_NEGINF and GSL_NAN no longer
cause divisions by zero during compilation.

** The gsl_interp_obj has been renamed to gsl_interp

** The poly_eval and pow_int functions have been moved from the
specfunc directory to the poly and sys directories.

** The Chebyshev functions are now available as an independent module
in their own directory.

** The error handling conventions have been unified across the
library.  This simplifies the use of the special functions.

** A full CBLAS implementation is now included for systems where ATLAS
has not been installed. The CBLAS library can also be used
independently of GSL.  The organisation of the BLAS directories has been
simplified.

** IEEE support for HPUX-11, NetBSD, Apple Darwin and OS/2 are now
included.

** The library now includes implementations of log1p, expm1, hypot,
acosh, asinh, atanh for platforms which do not provide them.

** The convention for alloc and set functions has changed so that they
are orthogonal. After allocating an object it is now necessary to
initialize it.

** There is a new module for estimating numerical derivatives of functions

** There is a new module for handling data with ntuples

** The histogram lookup functions are now optimized for the case of
uniform bins, and include an inline binary search for speed.

** The Chebyschev coefficients for the QAWO algorithm are now
precomputed in a table for efficiency, rather than being computed on
the fly.

** There are several new sorting functions for selecting the k-th
smallest or largest elements of a dataset.

** Iterator functions are now available for permutations,
gsl_permutation_next and gsl_permutation_prev.

** The function gsl_complex_xy has been renamed gsl_complex_rect

** The API for simulated annealing has been changed to support search
spaces in which the points cannot be represented as contiguous-memory
data structures.  gsl_siman_solve() now takes three extra arguments: a
copy constructor, a copy function and a destructor, allowing
gsl_siman_solve() to do its work with linked data structures.  If all
three of these function pointers are NULL, then the traditioanl
approach of using malloc(), memcpy(), and free() with the element size
is used.