Patch from bugtracking system of gphoto:
http://sourceforge.net/tracker/index.php?func=detail&aid=438155&group_id=8874&atid=108874

Summary:
gphoto2 leave DC240 in an unstable state
Fetching a list of all the files on the DC240 camera
leave the camera in an unstable state, making it hard
to shutdown.

Was applied in revision 1.4 to the ghoto2 source tree, plus an additional
patch, so use this version and not the patch from above.