samba20 - 2.0.x release version of samba, the UNIX SMB server
Samba 2.0.x (2.0.10) has functionality that is missing from the Samba 2.2.x
releases, one of which is an operational "share modes" directive.
According to the Samba mailing lists, there isn't a way in 2.2.x to do
what "share modes = No" did in 2.0.x, so revive an older, stable release
of Samba that does contain the missing functionality.  Setting the share
mode on a service is necessary only to get certain legacy MS-DOS
applications working in a multi-user environment.