initial import of gerbv.
Gerber Viewer (gerbv) is a viewer for Gerber files. Gerber files
are generated from PCB CAD system and sent to PCB manufacturers
as basis for the manufacturing process.

The different layers of the PCB are separated into different files.
gerbv can load all files at the same time, though it can not show
them at the same time. You have to browse through the different layers
with the radio buttons on the right side.