Initial import of ogle 0.8.2.
Ogle is a open source DVD player that supports DVD menus and
navigation. You should be able to play DVDs  directly from the
device, a mounted fs, an image file or as copy on you harddrive.
Keyboard commands for navigation in both player modes, if you
have input foucus in the video window.