Initial import of autoconf-2.52 (as autoconf-devel) into the Packages Collection.
Autoconf is an extensible package of m4 macros that produce shell
scripts to automatically configure software source code packages.
These scripts can adapt the packages to many kinds of UNIX-like
systems without manual user intervention.  Autoconf creates a
configuration script for a package from a template file that lists the
operating system features that the package can use, in the form of m4
macro calls.

Version 2.52 of autoconf is not entirely backwards compatible with
earlier versions, but is still needed to build some of the newer
packages. Import this package for just now as autoconf-devel.