Initial import of Erik E. Fair's shlock utility, with sources taken from NetBSD-current, into the NetBSD Packages Collection.
The shlock command can create or verify a lock file on behalf of a
shell or other script program.  When it attempts to create a lock
file, if one already exists, shlock verifies that it is or is not
valid.  If valid, shlock will exit with a non-zero exit code.  If
invalid, shlock will re- move the lock file, and create a new one.

shlock uses the rename(2) system call to make the final target lock
file, which is an atomic operation (i.e.  "dot locking", so named for
this mech- anism's original use for locking system mailboxes).  It
puts the process ID ("PID") from the command line into the requested
lock file.

shlock verifies that an extant lock file is still valid by using
kill(2) with a zero signal to check for the existence of the process
that holds the lock.

This package is only for Solaris and Linux platforms at present.