Apply QMAILQUEUE patch:
Bruce Guenter has written a patch which causes any program that would
run qmail-queue to look for an environment variable QMAILQUEUE.
If it is present, it is used in place of the string "bin/qmail-queue"
when running qmail-queue. This could be used, for example, to add a program
into the qmail-smtpd->qmail-queue pipeline that could do filtering,
rewrite broken headers, etc.

This does not break anything, and is required by some programs, for instance
qmail-scanner.

Bump PKGREVISION - we are at qmail-1.03nb2 now.