import of atlc-2.32
------

Atlc is a finite difference programme that is used to calculate the
properties of a two-conductor electrical transmission line of
arbitrary cross section. It is used whenever there are no analytical
formula known, yet you still require an answer. It can calculate:

  The impedance Zo (in Ohms)
  The capacitance per unit length (pF/m)
  The inductance per unit length (nF/m)
  The velocity of propogation v (m/s)
  The velocity factor, v/c, which is dimensionless.

A bitmap file (usually with the extension .bmp or .BMP) of the cross
section of the transmission line is drawn in a graphics package such
as The Gimp and then analyzed using Atlc.