Added conserver-7.1.3: allows multiple users to watch serial consoles
Conserver is an application that allows multiple users to watch a
serial console at the same time. It can log the data, allows users
to take write-access of a console (one at a time), and has a variety
of bells and whistles to accentuate that basic functionality.

The idea is that conserver will log all your serial traffic so you
can go back and review why something crashed, look at changes (if
done on the console), or tie the console logs into a monitoring
system (just watch the logfiles it creates).

With multi-user capabilities you can work on equipment with others,
mentor, train, etc.

It also does all that client-server stuff so that, assuming you
have a network connection, you can interact with any of the equipment
from home or wherever.

Submitted by Eric Schnoebelen <eric@cirr.com> in PR 15238,
DEINSTALL script submitted by dawszy@arhea.net in private mail.