Initial import of p11-2.10a, a PDP-11 emulator, into the NetBSD Packages Collection.
P11 is a PDP11 emulator, which currently emulates the following configuration:
        - a KDJ11A processor (including fpp)
        - one or more RL-controllers (RLV12) with up to 4 RL02 each
        - one or more RHV?? controllers with up to 8 RP06 each
        - an RK11-controller with 22-bit addressing (hey ever saw one?)
        - a MRV12 option (boot-rom/maintenance and display register)
        - one or more KL11A controllers with up to 4 serial lines each
        - a standard printer interface
        - a read only TM11 tape
        - a DEQNA ethernet adapter
        - a toy clock

On a 900MHz Athlon under FreeBSD 5.0 a 'make build; make installsrc' of
2.11BSD takes around 2:24 hours - this is several times faster than
a real KDJ11A (if /usr is mounted async,noaccesstime).
Floating point speed is in the order of 10 times the original.

P11 successfully runs RSX11M-PLUS, UN*X Version 5/6/7, 2.11BSD, RT-4 and XXDP.

With thanks to David Talmage for the nudge.