Initial import of opencm-0.1.0alpha16 into the Packages Collection.
Provided in PR 18064 by Julio Merino <jmmv@hispabsd.org>.

Modified by myself to include the buildlink2 files for boehm-gc,
ncurses, openssl, and the bsd.pkg.install.mk files, to use the
boehm-gc package, and not the included version, and to find one of the
help files appropriately.  I also fixes the CONF_FILES definition, and
tried to bring some sanity to the opencm version number.