Initial import of postgresql-autodoc 0.99
This is a utility which will run through PostgreSQL system tables and
returns HTML, DOT, and 2 styles of XML which describes the database.
The HTML is human readable (via webbrowser). The first style of XML is
actually the fileformat of Dia, a UML diagram tool. The second type of
XML is similar to the HTML but in the Docbook 4 format. It enables you
to mix in other docbook documentation via the XREFs, generating PDFs,
HTML, RTF, or other formatted documents. Between these tools and
JavaDoc with the appropriate XREFs, documentation about a project can
be generated quickly and be easily updatable yet have a very
professional look with some DSSSL work.