Initial import of Navi2ch 1.6.0.
Navi2ch is a viewer program dedicated to chatting in 2ch.net - the biggest,
most famous, influencial BBS in Japan.

Although this software is really well-written, It's worth noting that most
opinions/informations found in 2ch.net are next to senseless.  Thus, Good
NetBSD users are really encouraged not to devote themselves to 2ch! :-)