Update to 1.0beta3.  Changes: In 1.0beta3: This is the third beta for Speex, implementing what should be the last new features before 1.0. These new features are a new "ultra-wideband" mode for encoding at 32kHz (up to 48kHz) and an intensity stereo mode. Both of these are implemented to preserve backward and forward compatibility with other releases. This means that it is now possible to encode almost any wave file using Speex. Most of the work left for 1.0 is now quality improvements regarding VBR, perceptual enhancement and packet loss concealment.
In 1.0beta2:
This second beta release brings many, many optimizations to the
encoder and decoder. In some cases, the speedup is on the order of
50% to 100%. As part of the optimizations, some sections of Speex
have been written in assembly with SSE instructions which can be
enabled with the _USE_SSE flag when compiling with GCC. Finally,
the code has received a major cleanup.