Initial import of sysbuild-1.0. Description follows:
sysbuild is a script that simplifies the process of building NetBSD
releases and kernels as an unprivileged user. It should be clear enough
that it does not add any kind of magic to the NetBSD build system. It
just automates boring steps, like setup of permissions, directories,
execution of the build process, etc. It also provides enough
functionality to automatically update CVS source trees.

sysbuild can be easily used from command line or as a cron job, so
you can schedule most of the build jobs (really, it has been designed
to run perfectly from cron).

Approved by wiz.