Update cyrus-imapd to 2.1.11.
This includes the fix for the recent buffer overflow.

Note that IPv6 patch doesn't work, I've not had the time to look into this,
but felt that being a security fix I should get the update done.

Also note that if you define CYRUS_DB_RECOMMENDED in your build environment
it will build using the cyrus mailing list recommended defaults for the
database settings.